Refactor parserUtils class
